+++ jablib/src/main/java/org/jabref/logic/ai/util/MVStoreBase.java
@@ -12,6 +12,9 @@ import org.jspecify.annotations.NonNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+/// This is a base class for the AI artifacts (summaries, chat messages, etc.).
+/// It sets up and wraps an [MVStore], and also provides a way to load an *in-memory* storage in case there are problems while initializing the on-disk version.
+/// It also uses a [NotificationService] to notify the user about errors.
 public abstract class MVStoreBase implements AutoCloseable {
     private static final Logger LOGGER = LoggerFactory.getLogger(MVStoreBase.class);